requirements.txt
aiohttp>=3.8.3
aiosignal>=1.3.1
async-timeout>=4.0.2
attrs>=22.1.0
bleach>=5.0.1
cachetools>=5.2.0
certifi>=2022.9.24
charset-normalizer>=2.1.1
click>=8.1.3
click-log>=0.4.0
colorama>=0.4.6
dataclasses>=0.6
deprecation>=2.1.0
docutils>=0.19
dotty-dict>=1.3.1
Flask>=2.2.2
frozenlist>=1.3.3
geojson>=2.5.0
gitdb>=4.0.9
GitPython>=3.1.29
gotrue>=0.5.4
h11>=0.14.0
httpcore>=0.16.1
httpx>=0.23.1
idna>=3.4
importlib-metadata>=5.0.0
iniconfig>=1.1.1
invoke>=1.7.3
ipinfo>=4.4.2
itsdangerous>=2.1.2
jaraco.classes>=3.2.3
Jinja2>=3.1.2
keyring>=23.11.0
MarkupSafe>=2.1.1
more-itertools>=9.0.0
multidict>=6.0.2
newsapi-python>=0.2.6
packaging>=21.3
pkginfo>=1.8.3
pluggy>=1.0.0
postgrest-py>=0.10.3
pydantic>=1.10.2
Pygments>=2.13.0
pyowm>=3.3.0
pyparsing>=3.0.9
PySocks>=1.7.1
pytest>=7.2.0
python-dateutil>=2.8.2
python-dotenv>=0.21.0
python-gitlab>=3.11.0
python-semantic-release>=7.32.1
readme-renderer>=37.3
realtime>=0.0.5
requests>=2.28.1
requests-toolbelt>=0.10.1
rfc3986>=1.5.0
semver>=2.13.0
six>=1.16.0
smmap>=5.0.0
sniffio>=1.3.0
storage3>=0.3.5
supabase>=0.7.1
supafunc>=0.2.2
tomli>=2.0.1
tomlkit>=0.11.6
tqdm>=4.64.1
twine>=3.8.0
typing_extensions>=4.4.0
urllib3>=1.26.12
webassets>=2.0
webencodings>=0.5.1
websockets>=10.4
Werkzeug>=2.2.2
yarl>=1.8.1
zipp>=3.10.0
